Filling in ShiftStatus, Eligibility, adding more Scheduling classes
- completed methods for ShiftStatus state classes
- added pseudocode for Eligibility classes
- Added WorkDay, WorkWeek for containing time bounded information like
hours worked, daily open/close times, pay periods
- added isX checker methods in EmployeeType and Employee
- added a bunch of JavaDoc comments throughout

+++ src/main/java/users/EmployeeFactory.java
@@ -29,7 +29,7 @@ public class EmployeeFactory {
 								String name,
 								String email,
 								String type,
-								int seniority
+								Long   seniority
 								) {
 		if (FULL_TIME.equalsIgnoreCase(type)) {
 			return new FullTimeEmployee.Builder()
@@ -83,7 +83,7 @@ public class EmployeeFactory {
 						   .setName(name)
 						   .setEmail(email)
 						   .setType(EmployeeType.FULL_TIME)
-						   .setSeniority(0)
+						   .setSeniority(0L)
 						   .setElStrat(new FullTimeEligibility())
 						   .build();
 		}
@@ -92,7 +92,7 @@ public class EmployeeFactory {
 		 				   .setName(name)
 		 				   .setEmail(email)
 		 				   .setType(EmployeeType.PART_TIME)
-		 				   .setSeniority(0)
+		 				   .setSeniority(0L)
 						   .setElStrat(new PartTimeEligibility())
 		 				   .build();
 		}
@@ -101,7 +101,7 @@ public class EmployeeFactory {
 						 .setName(name)
 						 .setEmail(email)
 						 .setType(EmployeeType.CASUAL)
-						 .setSeniority(0)
+						 .setSeniority(0L)
 						 .setElStrat(new CasualEligibility())
 						 .build();
 		}
